backend/requirements.txt
django
djangorestframework
django-rest-swagger
ipython
django-money
psycopg2-binary

